More project reorganization.
- Renamed PlayWithMe/ to Classes/
- Created an extension to get the ItemRowView easier.
- Made a generic ItemRowView class to subclass for easier type casting.
