+++ Classes/Models/Chat.swift
