+++ Classes/Resources/Assets.xcassets/Contents.json
