+++ Classes/Resources/Assets.xcassets/chat_filled_icon.imageset/Contents.json
