+++ Classes/Resources/Assets.xcassets/chat_filled_icon.imageset/chat_filled_icon.png
