+++ Classes/Resources/Assets.xcassets/chat_filled_icon.imageset/chat_filled_icon@2x.png
