+++ Classes/Resources/Assets.xcassets/chat_icon.imageset/Contents.json
