+++ Classes/Resources/Assets.xcassets/chat_icon.imageset/chat_icon@2x.png
