+++ Classes/Resources/Assets.xcassets/chat_icon.imageset/chat_icon@3x.png
