+++ Classes/Resources/Base.lproj/ChatItemRowView.xib
