+++ Classes/Views/ChatItemRowView.swift
@@ -1,0 +1,15 @@
+//
+//  ChatItemRowView.swift
+//  PlayWithMe
+//
+//  Created by Alton Lau on 7/24/19.
+//  Copyright © 2019 Yelp Inc. All rights reserved.
+//
+
+import UIKit
+
+class ChatItemRowView: ItemRowView {
+  
+  @IBOutlet weak var subtitleLabel: UILabel!
+  
+}
