+++ Classes/Views/ItemRowView.swift
@@ -1,0 +1,16 @@
+//
+//  ItemRowView.swift
+//  PlayWithMe
+//
+//  Created by Alton Lau on 7/25/19.
+//  Copyright © 2019 Yelp Inc. All rights reserved.
+//
+
+import UIKit
+
+class ItemRowView: UIView {
+  
+  @IBOutlet weak var imageView: UIImageView!
+  @IBOutlet weak var titleLabel: UILabel!
+  
+}
